requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
djangorestframework==3.15.1
djoser==2.2.2
django-cors-headers==4.3.1
python-dotenv==1.0.1